Add TypeScript migration dashboard
As we convert parts of the codebase to TypeScript, we will want a way to
track progress. This commit adds a dashboard which displays all of the
files that we wish to convert to TypeScript and which files we've
already converted. The list of all possible files to convert has been
predetermined by using the `madge` package to start with all of the
entrypoints the build system uses to build the distributable version of
the extension and derive a dependency graph based on each entrypoint
(the files that the entrypoint imports, the files that the imports
import, etc). (The script to do this is included in this commit so it
can be re-run if needed.)

The dashboard is implemented as a separate React app. The CircleCI
configuration has been updated so that when a new commit is pushed, the
React app is built, and the built files are pushed to a separate repo
whose sole purpose is to serve the dashboard via GitHub Pages (this is
the same way that the Storybook works). All of the app code and script
to build the app are self-contained under
`development/ts-migration-dashboard`.